add support for spherical coordinates. To use spherical coodinates, set the number of dimensions to 3 (mandatory), set the "model" option to "sphere" and the radius option to the desired radius in units of light-microseconds. The coordinates are represented in cartesian space; I have a script to convert them to lat-long for easier viewing.

+++ p2psim/misc/vivalditest.C
@@ -328,9 +328,9 @@ VivaldiTest::print_all_loc()
   VivaldiNode::Coord vc;
   for (uint i = 0; i < n; i++) {
     vc = _all[i]->my_location();
-    printf ("%d at ", (int) _all[i]->ip());
+    printf ("%d at ",  _all[i]->ip());
     for (uint j = 0; j < vc._v.size(); j++)
-      printf ("%s%d", j ? "," : "", (int)vc._v[j]);
+      printf ("%s%f", j ? "," : "", vc._v[j]);
     if ((int)vc._ht)
       printf (",ht=%d", (int)vc._ht);
     printf (" with error %f\n", _all[i]->_last_error);
